feat: rework global cache and split responsibilities
