+++ server/api/repository/[owner]/[repo]/index.get.ts
@@ -1,4 +1,4 @@
-import { getRepoDetailsForRequest } from '~~/server/utils/shared-repo-data'
+import { getRepoDetailsForRequest } from '~~/server/utils/repo-sync-service'
 
 export default defineEventHandler(async (event) => {
   const { owner, repo } = getRepoParams(event)
