+++ server/utils/github.ts
@@ -31,26 +31,6 @@ export function updateRateLimitFromHeaders(headers: Headers, source: 'rest' | 'g
   }
 }
 
-/**
- * Invalidate the server-side issue detail cache after mutations.
- * Matches the key format used by defineCachedFunction in [number].get.ts
- */
-export async function invalidateIssueDetailCache(login: string, repo: string, issueNumber: number) {
-  const [owner, repoName] = repo.split('/')
-  if (!owner || !repoName) return
-  const cacheKey = `nitro:functions:issue-detail:${login}:${owner}/${repoName}#${issueNumber}.json`
-  await useStorage('cache').removeItem(cacheKey)
-}
-
-/**
- * Invalidate the server-side work item detail cache after mutations.
- * Matches the key format used by defineCachedFunction in work-items/[id].get.ts
- */
-export async function invalidateWorkItemDetailCache(login: string, owner: string, repo: string, id: string) {
-  const cacheKey = `nitro:functions:repo-work-item-detail:${login}:${owner}/${repo}:${id}.json`
-  await useStorage('cache').removeItem(cacheKey)
-}
-
 const GITHUB_API = 'https://api.github.com'
 
 export interface GitHubRequestOptions {
